Better handling of COORDS sets with certain commands. (#730)
* DRR - Cpptraj: Return a size_t instead of unsigned int; closer to the actual underlying type.

* DRR - Cpptraj: combinecrd now accepts wildcards in data set names and can use REF and TRAJ coords sets.

* DRR - Cpptraj: Allow TRAJ sets to be loaded as reference

* DRR - Cpptraj: Allow change command to access all COORDS type sets

* DRR - Cpptraj: Add test for change with COORDS set

* DRR - Cpptraj: Replace instances of FindCoordsSet with FindSetOfGroup; the former is only really appropriate for Analyses which may require creation of the default COORDS set. Ensure an output option is specified for PermuteDihedrals.

* DRR - Cpptraj: Revision bump for better COORDS set handling

+++ src/ParameterHolders.h
@@ -81,7 +81,7 @@ template <class T> class ParmHolder {
   public:
     ParmHolder() {}
     void clear() { bpmap_.clear(); }
-    unsigned int size() const { return bpmap_.size(); }
+    size_t size() const { return bpmap_.size(); }
 /*
     static inline void PrintTypes(AtomTypeHolder const& types) {
       for (AtomTypeHolder::const_iterator it = types.begin(); it != types.end(); ++it)
